congressMCP/requirements.txt
Package            Version
------------------ ---------
annotated-types    0.7.0
anyio              4.9.0
beautifulsoup4     4.13.4
cachetools         5.5.2
certifi            2025.4.26
charset-normalizer 3.4.2
click              8.2.1
distro             1.9.0
google-auth        2.40.3
google-genai       1.20.0
h11                0.16.0
httpcore           1.0.9
httpx              0.28.1
httpx-sse          0.4.0
idna               3.10
jiter              0.10.0
markdown-it-py     3.0.0
mcp                1.9.4
mdurl              0.1.2
openai             1.86.0
pip                25.1.1
pyasn1             0.6.1
pyasn1_modules     0.4.2
pydantic           2.11.6
pydantic_core      2.33.2
pydantic-settings  2.9.1
Pygments           2.19.1
python-dotenv      1.1.0
python-multipart   0.0.20
PyYAML             6.0.2
requests           2.32.4
rich               14.0.0
rsa                4.9.1
setuptools         65.5.0
shellingham        1.5.4
sniffio            1.3.1
soupsieve          2.7
sse-starlette      2.3.6
starlette          0.47.0
tqdm               4.67.1
typer              0.16.0
typing_extensions  4.14.0
typing-inspection  0.4.1
urllib3            2.4.0
uvicorn            0.34.3
websockets         15.0.1
